+++ src/format/reducer_conditions.js
@@ -9,6 +9,7 @@ import unwrap_field from '../utils/unwrap_field.js';
 /**
  * @typedef {import('sql-template-tag').Sql} Sql
  * @typedef {import('../index.js').default} Dare
+ * @typedef {import('../index.js').Engine} Engine
  */
 /* eslint-enable jsdoc/valid-types */
 
@@ -126,6 +127,8 @@ function prepCondition({
 	conditional_operators_in_value,
 	dareInstance,
 }) {
+	const {engine} = dareInstance;
+
 	// Does it have a negative comparison operator?
 	const negate = operators?.includes('-');
 
@@ -144,13 +147,20 @@ function prepCondition({
 
 	if (isFullText) {
 		// Join the fields
-		const sql_field = join(
-			sql_fields.map(({sql}) => sql),
-			', '
-		);
+		const sql_field_array = sql_fields.map(({sql}) => sql);
 
-		// Full Text
-		return SQL`${NOT}MATCH(${sql_field}) AGAINST(${dareInstance.fulltextParser(value)} IN BOOLEAN MODE)`;
+		const IS_POSTGRES = dareInstance.engine.startsWith('postgres');
+
+		if (IS_POSTGRES) {
+			const field =
+				sql_field_array.length === 1
+					? sql_field_array.at(0)
+					: SQL`TO_TSVECTOR(${join(sql_field_array, " || ' ' || ")})`;
+			return SQL`${NOT}${field} @@ to_tsquery('english', ${dareInstance.fulltextParser(value)})`;
+		}
+
+		// Default: MySQL Full Text
+		return SQL`${NOT}MATCH(${join(sql_field_array, ', ')}) AGAINST(${dareInstance.fulltextParser(value)} IN BOOLEAN MODE)`;
 	} else if (sql_fields.length > 1) {
 		/*
 		 * Is the field an array of field names?
@@ -195,7 +205,7 @@ function prepCondition({
 		!Array.isArray(value)
 	) {
 		// Loop through the object and create the sql_field
-		const sql_fields = json_contains(sql_field, value);
+		const sql_fields = json_contains({sql_field, value, engine});
 
 		// Return a single or a wrapped group
 		return SQL`${NOT}(${join(
@@ -206,6 +216,7 @@ function prepCondition({
 					conditional_operators_in_value,
 					operators,
 					type,
+					engine,
 				})
 			),
 			' AND '
@@ -219,6 +230,7 @@ function prepCondition({
 		operators,
 		// Treat json as text
 		type: type === 'json' ? 'text' : type,
+		engine,
 	});
 }
 
@@ -230,6 +242,7 @@ function prepCondition({
  * @param {string|null} params.conditional_operators_in_value - Allowable conditional operators in value
  * @param {string|null} params.operators - Operators
  * @param {string|null} params.type - Type
+ * @param {Engine} params.engine - DB Engine
  * @returns {Sql} SQL condition
  */
 function sqlCondition({
@@ -238,7 +251,10 @@ function sqlCondition({
 	conditional_operators_in_value,
 	operators,
 	type,
+	engine,
 }) {
+	const IS_POSTGRES = engine.startsWith('postgres');
+
 	// Does it have a negative comparison operator?
 	const negate = operators?.includes('-');
 
@@ -266,6 +282,8 @@ function sqlCondition({
 	// Conditional JSON Quote
 	const quote =
 		type === 'json' ? a => (typeof a === 'string' ? `"${a}"` : a) : a => a;
+
+	const LIKE = raw(IS_POSTGRES ? 'ILIKE' : 'LIKE');
 
 	/*
 	 * Range
@@ -304,7 +322,7 @@ function sqlCondition({
 		allow_conditional_negate_operator_in_value &&
 		value[0] === '!'
 	) {
-		return SQL`${sql_field} NOT LIKE ${value.slice(1)}`;
+		return SQL`${sql_field} NOT ${LIKE} ${value.slice(1)}`;
 	}
 
 	// String partial match
@@ -313,7 +331,9 @@ function sqlCondition({
 		(isLikey ||
 			(allow_conditional_likey_operator_in_value && value.match('%')))
 	) {
-		return SQL`${sql_field} ${NOT}LIKE ${quote(value)}`;
+		const strValue = !IS_POSTGRES ? quote(value) : value;
+
+		return SQL`${sql_field} ${NOT}${LIKE} ${strValue}`;
 	}
 
 	// Null
@@ -360,12 +380,10 @@ function sqlCondition({
 
 		// Use the `IN(...)` for items which can be grouped...
 		if (filteredValue.length) {
-			const items =
-				process.env.MYSQL_VERSION?.startsWith('8') ||
-				process.env.MYSQL_VERSION?.startsWith('5.6')
-					? filteredValue
-					: filteredValue.map(quote);
-			conds.push(SQL`${sql_field} ${NOT}IN (${items})`);
+			const items = engine.startsWith('mysql:5.7')
+				? filteredValue.map(quote)
+				: filteredValue;
+			conds.push(SQL`${sql_field} ${NOT}IN (${join(items)})`);
 		}
 
 		// Other Values which can't be grouped ...
@@ -377,6 +395,7 @@ function sqlCondition({
 					operators,
 					conditional_operators_in_value,
 					type,
+					engine,
 				})
 			)
 		);
@@ -386,25 +405,49 @@ function sqlCondition({
 			? conds.at(0)
 			: SQL`(${join(conds, negate ? ' AND ' : ' OR ')})`;
 	} else {
+		if (
+			IS_POSTGRES &&
+			type === 'json' &&
+			(typeof value === 'boolean' || typeof value === 'number')
+		) {
+			value = String(value);
+		}
+
 		return SQL`${sql_field} ${raw(negate ? '!' : '')}= ${value}`;
 	}
 }
 
 /**
  * JSON Contains
- * @param {Sql} sql_field - SQL Field
- * @param {any} value - Value
- * @param {string} [path] - Path
- * @param {string} [operators] - Operators
+ * @param {object} params - Params
+ * @param {Sql} params.sql_field - SQL Field
+ * @param {any} params.value - Value
+ * @param {string} [params.path] - Path
+ * @param {string} [params.operators] - Operators
+ * @param {string} [params.engine] - Engine
  * @returns {Array<{sql: Sql, value: any, operators: string}>} SQL conditions
  */
-function json_contains(sql_field, value, path = '$', operators = '') {
+function json_contains({
+	sql_field,
+	value,
+	path = null,
+	operators = '',
+	engine,
+}) {
+	const IS_POSTGRES = engine.startsWith('postgres');
+
+	if (!path && !IS_POSTGRES) {
+		path = '$';
+	}
+
 	const conds = [];
 
 	if (typeof value !== 'object' || value === null || Array.isArray(value)) {
+		const separator = IS_POSTGRES ? '->>' : '->';
+
 		return [
 			{
-				sql: SQL`${sql_field}->${path}`,
+				sql: SQL`${sql_field}${raw(separator)}${path}`,
 				value,
 				operators,
 			},
@@ -414,12 +457,13 @@ function json_contains(sql_field, value, path = '$', operators = '') {
 	for (const key in value) {
 		const {operators: newOperators, rootKey} = stripKey(key);
 		conds.push(
-			...json_contains(
+			...json_contains({
 				sql_field,
-				value[key],
-				`${path}.${rootKey}`,
-				operators + newOperators
-			)
+				value: value[key],
+				path: [path, rootKey].filter(Boolean).join('.'),
+				operators: operators + newOperators,
+				engine,
+			})
 		);
 	}